+++ satella/coding/structures/__init__.py
@@ -1,5 +1,6 @@
 from .dictionaries import DictObject, apply_dict_object, DictionaryView, TwoWayDictionary, \
-    DirtyDict, KeyAwareDefaultDict, ExpiringEntryDict, SelfCleaningDefaultDict
+    DirtyDict, KeyAwareDefaultDict, ExpiringEntryDict, SelfCleaningDefaultDict, \
+    CacheDict
 from .hashable_objects import HashableWrapper
 from .immutable import Immutable, frozendict
 from .proxy import Proxy
@@ -11,6 +12,7 @@ from .heaps import Heap, SetHeap, TimeBasedHeap, TimeBasedSetHeap
 from .typednamedtuple import typednamedtuple
 
 __all__ = [
+    'CacheDict',
     'KeyAwareDefaultDict',
     'Proxy',
     'ReprableMixin',