+++ satella/coding/structures/dictionaries/__init__.py
@@ -1,3 +1,4 @@
 from .objects import DirtyDict, DictionaryView, KeyAwareDefaultDict, TwoWayDictionary
 from .dict_object import apply_dict_object, DictObject
 from .expiring import ExpiringEntryDict, SelfCleaningDefaultDict
+from .cache_dict import CacheDict