+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.8.8
 
+* special-cased passing a DictObject to another DictObject