+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.8.8
 
 * special-cased passing a DictObject to another DictObject
+* added str and repr to DictObject