+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.8.9
+
+* added `CacheDict`