+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-# v2.8.9
+# v2.8.10
 
-* added `CacheDict`
+